+++ pkg/store/labelpb/label.go
@@ -73,6 +73,25 @@ func ZLabelSetsToPromLabelSets(lss ...ZLabelSet) []labels.Labels {
 	return res
 }
 
+// ZLabelSetsFromPromLabels converts []labels.labels to []labelpb.ZLabelSet.
+func ZLabelSetsFromPromLabels(lss ...labels.Labels) []ZLabelSet {
+	sets := make([]ZLabelSet, 0, len(lss))
+	for _, ls := range lss {
+		set := ZLabelSet{
+			Labels: make([]ZLabel, 0, len(ls)),
+		}
+		for _, lbl := range ls {
+			set.Labels = append(set.Labels, ZLabel{
+				Name:  lbl.Name,
+				Value: lbl.Value,
+			})
+		}
+		sets = append(sets, set)
+	}
+
+	return sets
+}
+
 // ZLabel is a Label (also easily transformable to Prometheus labels.Labels) that can be unmarshalled from protobuf
 // reusing the same memory address for string bytes.
 // NOTE: While unmarshalling it uses exactly same bytes that were allocated for protobuf. This mean that *whole* protobuf